+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/properties/mods/pehuki/MPRScalePehkui.java
@@ -13,6 +13,7 @@ import insane96mcp.mobspropertiesrandomness.data.json.util.modifiable.MPRRange;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraftforge.fml.ModList;
+import org.apache.logging.log4j.util.TriConsumer;
 import virtuoel.pehkui.api.ScaleData;
 import virtuoel.pehkui.api.ScaleRegistries;
 import virtuoel.pehkui.api.ScaleType;
@@ -27,25 +28,29 @@ public class MPRScalePehkui implements IMPRObject {
     @SerializedName("scale_types")
     public List<String> scaleTypes;
 
-    public MPRScalePehkui(MPRRange scale, List<String> scaleTypes) {
+    public Operation operation;
+
+    public MPRScalePehkui(MPRRange scale, List<String> scaleTypes, Operation operation) {
         this.scale = scale;
         this.scaleTypes = scaleTypes;
+        this.operation = operation;
     }
 
     @Override
     public void validate() throws JsonValidationException {
         if (this.scale == null)
-            throw new JsonValidationException("scale missing from MPRScalePehkui");
+            throw new JsonValidationException("scale missing from ScalePehkui");
         if (this.scaleTypes == null || this.scaleTypes.isEmpty())
-            throw new JsonValidationException("scale_types missing or empty from MPRScalePehkui");
+            throw new JsonValidationException("scale_types missing or empty from ScalePehkui");
+        if (this.operation == null)
+            throw new JsonValidationException("operation missing from ScalePehkui");
     }
 
     public void apply(LivingEntity entity) {
-        float scale = this.scale.getFloat(entity, entity.level);
         for (String scaleType : this.scaleTypes) {
             ScaleType type = ScaleRegistries.SCALE_TYPES.get(new ResourceLocation(scaleType));
             ScaleData scaleData = type.getScaleData(entity);
-            scaleData.setScale(scale);
+            this.operation.applyScale(scaleData, this.scale, entity);
         }
     }
 
@@ -58,7 +63,29 @@ public class MPRScalePehkui implements IMPRObject {
             if (!ModList.get().isLoaded("pehkui"))
                 throw new JsonParseException("Pehkui is not present. This object can't be used: %s.".formatted(json));
 
-            return new MPRScalePehkui(context.deserialize(json.getAsJsonObject().get("scale"), MPRRange.class), context.deserialize(json.getAsJsonObject().get("scale_types"), STRING_LIST_TYPE));
+            return new MPRScalePehkui(context.deserialize(json.getAsJsonObject().get("scale"), MPRRange.class), context.deserialize(json.getAsJsonObject().get("scale_types"), STRING_LIST_TYPE), context.deserialize(json.getAsJsonObject().get("operation"), Operation.class));
+        }
+    }
+
+    enum Operation {
+        @SerializedName("set")
+        SET((scaleData, mprRange, entity) ->
+                scaleData.setScale(mprRange.getFloat(entity, entity.level))),
+        @SerializedName("add")
+        ADD((scaleData, mprRange, entity) ->
+                scaleData.setScale(scaleData.getScale() + mprRange.getFloat(entity, entity.level))),
+        @SerializedName("multiply")
+        MULTIPLY((scaleData, mprRange, entity) ->
+                scaleData.setScale(scaleData.getScale() * mprRange.getFloat(entity, entity.level)));
+
+        final TriConsumer<ScaleData, MPRRange, LivingEntity> apply;
+
+        Operation(TriConsumer<ScaleData, MPRRange, LivingEntity> apply) {
+            this.apply = apply;
+        }
+
+        public void applyScale(ScaleData scaleData, MPRRange mprRange, LivingEntity entity) {
+            this.apply.accept(scaleData, mprRange, entity);
         }
     }
 }